requirements.txt
streamlit
litellm
requests
python-multipart
python-dotenv